Thunks: Adds libOpenCL
This takes a bit of faff due to all of the different style of callbacks
that it does. These have different lifetimes depending on the object
which means that a generator for these can't really exist.

Depends on the previous PRs to be merged before use.

Needs functional testing which I'll open a PR after this.