Dispatcher: Signal handling changes.
Probably best to view this one with whitespace disabled.

Minor things:
- Calculate eflags in the context when going in to a signal
- Enable SS uc_flags. We save and restore these so it should be set
- On x86_64 clear RAX going in to a signal.
  - Seems to be a workaround for buggy variadic C applications.

More Major work:
- x86_64 now calculates mcontext and siginfo at all times
  - This matches Linux kernel behaviour where this is always calculated
- x86 now calculates mcontext at all times.
  - This matches Linux kernel behaviour where this is always calculated
  - siginfo is calculated only if SA_SIGINFO is set

FEX host behaviour change:

We now take a stack slot to store our context pointer on it. This will
be necessary in a commit after this where we change how signal returns
work. This doesn't solve the problem that long-jumping from a signal
handler will burn our host stack space, but if we need to, we have
enough minimum stack space remaining that we could store the whole
/host/ stack space on the guest stack.

Since the signal stack data is a handshake with the kernel with how signals
return, this is safe to do. Also we never supported applications setting
up a kernel stack frame and calling rt_sigreturn anyway.