Syscalls: Adds support for rt_sigreturn and sigreturn
By utilizing the stack unwinding using the guest stack space, we can now
implement these directly.

With all of this work in place, libunwind now successfully unwinds a
stack frame from a signal!

It needed to find the appropriate signal frame to know that it is in a signal stack to unwind accordingly.

Before:
```
*** signal 11
Register dump:

<... Register dump snipped ...>

Backtrace:
/mnt/Work/Projects/work/glibc-build/debug/libSegFault.so(+0x37b4)[0x7f245c9537b4]
[0x7f245c996000]
```

After:
```
*** signal 11
Register dump:

<... Register dump snipped ...>

Backtrace:
/mnt/Work/Work/Unicorn_Tests/ELFLoader/Tests/invalid_int3(+0x1130)[0x5589d3060130]
/mnt/Work/Work/Unicorn_Tests/ELFLoader/Tests/invalid_int3(+0x1146)[0x5589d3060146]
/lib/x86_64-linux-gnu/libc.so.6(+0x29d90)[0x7fffe25d7d90]
/lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0x80)[0x7fffe25d7e40]
/mnt/Work/Work/Unicorn_Tests/ELFLoader/Tests/invalid_int3(+0x1065)[0x5589d3060065]
```