Thunks/gen: Add support for assisted automatic repacking of structs
Repacking is fully automatic for structures composed of simple types
with compatible ABI across guest/host architectures.

To repack structures with pointers, the custom_repack annotation can be
used. Repacking will still be automated then (with the pointer being
extended from guest-size to host-size), but a call to a user-provided
function is emitted to handle architecture differences.